invert Y axis, fix some rounding errors in CPU graph
Inverting Y axis helps for recognition of boundary cases - borders of
my monitor are black, and since background is also black you can't
reliably see how close you are to the maximum (it blends into the
border). When inverted, the border of the chart is close to other
windows which are usually not black.

CPU chart had "jumping maximum" issue - glibtop returns some elapsed
CPU time values, and total amount is not the same on each update. I added
division by total amount of ticks and makes sure that the maximum
value is the same for all time points.